[merb_datamapper] Fixed session configuration.
Because DataMapperSession was always required, the sessions table would always be created during an auto_migrate!.
And the check to see if DataMapperSession should be required was using the wrong configuration.
Also, due to the ordering of storage_names[default_repository_name] and the definition of self.default_repository_name, the configured default_repository_name was never used.

These have been fixed.